user auth

+++ app/Http/Controllers/TasksController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Task;
 
+
 class TasksController extends Controller
 {
     /**
@@ -14,8 +15,20 @@ class TasksController extends Controller
      */
     public function index()
     {
-        $tasks = Task::orderBy('id','desc')->paginate(25);
-        return view('tasks.index',['tasks'=>$tasks]);
+        // $tasks = Task::orderBy('id','desc')->paginate(25);
+        // return view('tasks.index',['tasks'=>$tasks]);
+        
+        $data = [];
+        if(\Auth::check()) {
+            $user = \Auth::user();
+            $tasks = $user->tasks()->orderBy('created_at','desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        return view('tasks.index',$data);
+        
     }
 
     /**
@@ -43,10 +56,19 @@ class TasksController extends Controller
             'status' => 'required|max:10',
         ]);
         
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        // $task = new Task;
+        // $task->content = $request->content;
+        // $task->status = $request->status;
+        // //追記
+        // $task->user_id = auth()->id();
+        // // dd($request->all());
+        // $task->save();
+        
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status,
+            'user_id' => auth()->id(),
+        ]);
         
         return redirect('/');
     }
@@ -59,9 +81,13 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::findOrFail($id);
+        if(\Auth::check()) {
+            $task = Task::findOrFail($id);
+            return view('tasks.show',['task'=>$task,]);
+        } else {
+            return redirect('/');
+        }
         
-        return view('tasks.show',['task'=>$task,]);
     }
 
     /**
@@ -110,7 +136,10 @@ class TasksController extends Controller
     public function destroy($id)
     {
         $task = Task::findOrFail($id);
-        $task->delete();
+        // $task->delete();
+        if(\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
         
         return redirect('/');
     }
